feat: convert action files to typescript and use new redux actions
